feat: new custom hook: useWeather feat: new component: WeatherEasterEgg chore: moved dark default flares to default-flares
