+++ src/components/content/heroes/aurora/default-flares.tsx
@@ -37,3 +37,29 @@ export const DefaultFlares = () => {
     </>
   );
 };
+
+export const DarkDefaultFlares = () => {
+  return (
+    <>
+      <Flare
+        stepSize={20}
+        flareType={FlareType.DARK}
+        x={'50vw'}
+        y={'50vw'}
+        size={150}
+        rotation={30}
+        animationOffset={3}
+      />
+      <Flare
+        stepSize={0}
+        flareType={FlareType.DARK}
+        x={'70vw'}
+        y={'300px'}
+        size={100}
+        rotation={80}
+        animationOffset={14}
+      />
+      <DefaultFlares />
+    </>
+  );
+};
